领取任务接口

+++ renrenentity/src/main/java/com/renrentui/renrenentity/req/TaskDetailReq.java
@@ -6,18 +6,18 @@ package com.renrentui.renrenentity.req;
  *
  */
 public class TaskDetailReq {
-	private int  userId;
-	private int  taskId;
-	public int getUserId() {
+	private Long  userId;
+	private Long  taskId;
+	public Long getUserId() {
 		return userId;
 	}
-	public void setUserId(int userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
-	public int getTaskId() {
+	public Long getTaskId() {
 		return taskId;
 	}
-	public void setTaskId(int taskId) {
+	public void setTaskId(Long taskId) {
 		this.taskId = taskId;
 	}
 }
